requirements.txt
# Core dependencies
torch>=2.0.0,<3.0.0
transformers>=4.30.0,<5.0.0
datasets>=2.0.0,<3.0.0
openai>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
tqdm>=4.65.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0
tabulate>=0.9.0,<1.0.0


# Optional: For HarmFormer model loading
# huggingface_hub>=0.16.0,<1.0.0

# Optional: For Perspective API baseline
# google-api-python-client>=2.100.0,<3.0.0

# Optional: For Llama Guard baseline (already covered by torch + transformers)
# accelerate>=0.20.0,<1.0.0
